Refactored the test
